requirements.txt
wheel
openai-whisper
SpeechRecognition
openai
gtts
pygame
pywhatkit
wikipedia
pyjokes
pyttsx3
pyautogui
python-decouple
pyyaml